+++ index.js
@@ -51,6 +51,7 @@ module.exports = {
       multiline: 'below',
     }],
     'vue/first-attribute-linebreak': 'error',
+    'vue/multi-word-component-names': 'off',
     'vue/no-multiple-template-root': 'off',
     'vue/no-reserved-props': ['error', {
       vueVersion: 3,